+++ Modules/Intent.Modules.Common.CSharp/Templates/ICSharpTemplate.cs
@@ -1,6 +1,7 @@
 ﻿using Intent.Metadata.Models;
 using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.FileBuilders;
+using Intent.Modules.Common.CSharp.FactoryExtensions;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.VisualStudio;
 
@@ -19,6 +20,11 @@ public interface ICSharpTemplate : IIntentTemplate, IHasNugetDependencies, IHasA
     string GetFullyQualifiedTypeName(string templateId, TemplateDiscoveryOptions options = null);
     string GetFullyQualifiedTypeName(string templateId, IMetadataModel model, TemplateDiscoveryOptions options = null);
     string GetFullyQualifiedTypeName(string templateId, string modelId, TemplateDiscoveryOptions options = null);
+
+    void AddKnownType(string fullyQualifiedTypeName)    
+    {
+        CSharpTypesCache.AddKnownType(fullyQualifiedTypeName);
+    }
 }
 
 public interface ICSharpFileBuilderTemplate : IFileBuilderTemplate, ICSharpTemplate