+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -3,6 +3,7 @@
 - New Feature: `CSharpType` to represent a C# Type that can be more easily mutated and reliably inspected for the type it is. It has a parser to convert `string` into a `CSharpType`.
 - New Feature: `CSharpVariableDeclaration` and `CSharpDeclarationExpression` to represent variable declarations in C# like `var myVar` and `var (prop1, prop2, prop3)`.
 - Improvement: `CSharpClass`, `CSharpClassMethod`, `CSharpInterface` and `CSharpInterfaceMethod` are able to accomodate return types in the form of `CSharpType`.
+- Improvement: Support for `ValueTask` as a `CSharpType` on Methods.
 
 ### Version 3.6.4